Adding R code.
Note that the downloaddata.py file in the r directory is a copy of the
one in the python directory. The R download script invokes the Python
script. Ideally a soft link to the script in the Python directory
would work. Unfortunatly, git on windows appears to replace soft links
with an ASCII file containing the full path to the original file. So I
ended up just copying the file over.